+++ tests/adapter/mongo/test_evaluated_variants.py
@@ -183,7 +183,6 @@ def test_get_ranked_and_comment_two(real_variant_database):
 
 
 def test_evaluated_variants(case_obj, institute_obj, user_obj, real_variant_database):
-
     adapter = real_variant_database
     case_id = case_obj["_id"]
     institute_id = case_obj["owner"]