+++ tests/adapter/mongo/test_variant_events.py
@@ -75,7 +75,6 @@ def test_matching_manual_rank(
 
 
 def test_mark_causative(adapter, institute_obj, case_obj, user_obj, variant_obj):
-
     # GIVEN a populated database with variants
     adapter.case_collection.insert_one(case_obj)
     adapter.institute_collection.insert_one(institute_obj)
@@ -109,7 +108,6 @@ def test_mark_causative(adapter, institute_obj, case_obj, user_obj, variant_obj)
 
 
 def test_unmark_causative(adapter, institute_obj, case_obj, user_obj, variant_obj):
-
     ## GIVEN a adapter with a variant that is marked causative
     adapter.case_collection.insert_one(case_obj)
     adapter.institute_collection.insert_one(institute_obj)
@@ -147,7 +145,6 @@ def test_unmark_causative(adapter, institute_obj, case_obj, user_obj, variant_ob
 
 
 def test_mark_partial_causative(adapter, institute_obj, case_obj, user_obj, variant_obj):
-
     # GIVEN a populated database with variants
     adapter.case_collection.insert_one(case_obj)
     adapter.institute_collection.insert_one(institute_obj)
@@ -184,7 +181,6 @@ def test_mark_partial_causative(adapter, institute_obj, case_obj, user_obj, vari
 
 
 def test_unmark_partial_causative(adapter, institute_obj, case_obj, user_obj, variant_obj):
-
     # GIVEN a populated database with variants
     adapter.case_collection.insert_one(case_obj)
     adapter.institute_collection.insert_one(institute_obj)
@@ -230,7 +226,6 @@ def test_unmark_partial_causative(adapter, institute_obj, case_obj, user_obj, va
 
 
 def test_order_verification(adapter, institute_obj, case_obj, user_obj, variant_obj):
-
     # GIVEN a populated database with variants
     adapter.case_collection.insert_one(case_obj)
     adapter.institute_collection.insert_one(institute_obj)
@@ -269,7 +264,6 @@ def test_order_verification(adapter, institute_obj, case_obj, user_obj, variant_
 
 
 def test_cancel_verification(adapter, institute_obj, case_obj, user_obj, variant_obj):
-
     # GIVEN a populated database with a variant that has sanger ordered
     adapter.case_collection.insert_one(case_obj)
     adapter.institute_collection.insert_one(institute_obj)
@@ -309,7 +303,6 @@ def test_cancel_verification(adapter, institute_obj, case_obj, user_obj, variant
 
 
 def test_dismiss_variant(adapter, institute_obj, case_obj, user_obj, variant_obj):
-
     # GIVEN a variant db with at least one variant, and no events
     adapter.case_collection.insert_one(case_obj)
     adapter.institute_collection.insert_one(institute_obj)
@@ -347,7 +340,6 @@ def test_dismiss_variant(adapter, institute_obj, case_obj, user_obj, variant_obj
 
 
 def test_matching_dismissed_variant(adapter, institute_obj, case_obj, user_obj, variant_obj):
-
     # GIVEN a variant db with at least one variant, and no events
     adapter.case_collection.insert_one(case_obj)
     adapter.institute_collection.insert_one(institute_obj)
@@ -408,7 +400,6 @@ def test_matching_dismissed_variant(adapter, institute_obj, case_obj, user_obj, 
 
 
 def test_update_cancer_tier(adapter, institute_obj, case_obj, user_obj, variant_obj):
-
     # GIVEN a variant db with at least one variant, and no events
     adapter.case_collection.insert_one(case_obj)
     adapter.institute_collection.insert_one(institute_obj)
@@ -445,7 +436,6 @@ def test_update_cancer_tier(adapter, institute_obj, case_obj, user_obj, variant_
 
 
 def test_update_manual_rank(adapter, institute_obj, case_obj, user_obj, variant_obj):
-
     # GIVEN a variant db with at least one variant, and no events
     adapter.case_collection.insert_one(case_obj)
     adapter.institute_collection.insert_one(institute_obj)