+++ tests/server/blueprints/variants/test_variants_views.py
@@ -10,7 +10,6 @@ from scout.server.extensions import store
 def test_variants_clinical_filter(
     app, real_variant_database, institute_obj, case_obj, mocker, mock_redirect
 ):
-
     mocker.patch("scout.server.blueprints.variants.views.redirect", return_value=mock_redirect)
 
     # GIVEN a variant without clinVar annotations. Also needs to be present in affected individual, have gene id in panel,